mapeo de la entidad servicio
correccion del tipo del id

clases repository e implement

se agrego el atributo estado

se indico la clave primaria

se modifico la vista del id en el formulario

+++ TP2_Grupo21/src/main/java/ar/edu/unju/fi/service/IServicioService.java
@@ -12,6 +12,6 @@ public interface IServicioService {
 	void listar();
 	void modificar(Servicio servicio);
 	void eliminar(Servicio servicio);
-	boolean buscar(int id);
-	Servicio recuperar(int id);
+	boolean buscar(Long id);
+	Servicio recuperar(Long id);
 }
